refactor: move text_cleanup to wiki_parsing
